[refactor] : 폴더구조 변경
- domain 폴더 안에서 page/utils/service 구분해서 파일 정리
- class가 아닌 function 으로 이루어진경우 대문자시작 => 카멜 케이스로 변경
- types 파일 변경
- 상태 관리가 필요없는 경우 class에서 function으로 변경
