+++ index.html
@@ -11,8 +11,5 @@
 
   <body>
     <script src="./src/main.js" type="module"></script>
-
-    <!-- GNB -->
-    <!-- header 들어오는 자리 -->
   </body>
 </html>
