+++ src/types/domain/GetInputType.ts
