+++ src/types/domain/LoadFoodListType.ts
@@ -1,0 +1,3 @@
+export interface LoadFoodListType {
+  title: string;
+}
