--- src/types/pages/FoodListPageType.ts
@@ -1,13 +1,0 @@
-export interface FoodListPageType {
-  title: string;
-  iconButton: Element | null;
-}
-
-export interface LoadFoodListType {
-  title: string;
-}
-
-export interface LoadHeaderType {
-  title: string;
-  iconButton: Element | null;
-}
