+++ templates/index.html
@@ -8,8 +8,6 @@
   </head>
 
   <body>
-    <div class="alert-container"></div>
-
     <!-- GNB -->
     <header class="gnb">
       <h1 class="gnb__title text-title">점심 뭐 먹지</h1>
@@ -242,5 +240,4 @@
       </div>
     </main>
   </body>
-
 </html>
